requirements.txt
requests==2.32.3
flask==3.1.0
flask-cors==5.0.0
pdfplumber==0.11.4
PyPDF2==3.0.1
python-dotenv==1.0.1
Werkzeug==3.1.3
